feat(github-pages): copy app directory to new docs directory
There doesn't seem to be a way currently to set a custom directory in GitHub Pages for it to display that so I am using the `/docs` directory instead.
On every pull in this repo I will manually replace (or maybe try to create a ci script for that in the future) the new contents in the `/app` directory with the contents of the `/docs` directory.